test: add a11y tests to all pages
